requirements.txt
aiogram>=3.10,<4.0
aiohttp>=3.9,<4.0
SQLAlchemy>=2.0,<3.0
aiosqlite>=0.20,<1.0
python-dotenv>=1.0,<2.0
